Simplify the UIEffects::GetCentre function.
This function operates on a QRect, whose coordinates are all of type
qreal (aka double).  Change the function to keep everything as double
precision, instead of converting to/from floats and processing as
floating point precision.

+++ mythtv/libs/libmythui/mythuianimation.h
@@ -22,16 +22,16 @@ class UIEffects
 
     QPointF GetCentre(const QRect rect, int xoff, int yoff) const
     {
-        float x = static_cast<float>(xoff) + static_cast<float>(rect.left());
-        float y = static_cast<float>(yoff) + static_cast<float>(rect.top());
+        qreal x = static_cast<qreal>(xoff) + rect.left();
+        qreal y = static_cast<qreal>(yoff) + rect.top();
         if (Middle == m_centre || Top == m_centre || Bottom == m_centre)
-            x += static_cast<float>(rect.width()) / 2.0F;
+            x += rect.width() / 2.0;
         if (Middle == m_centre || Left == m_centre || Right == m_centre)
-            y += static_cast<float>(rect.height()) / 2.0F;
+            y += rect.height() / 2.0;
         if (Right == m_centre || TopRight == m_centre || BottomRight == m_centre)
-            x += static_cast<float>(rect.width());
+            x += rect.width();
         if (Bottom == m_centre || BottomLeft == m_centre || BottomRight == m_centre)
-            y += static_cast<float>(rect.height());
+            y += rect.height();
         return {x, y};
     }
